Move defaultMain and startWithCliArgs to Termonad.Startup module
