+++ src/Termonad.hs
@@ -5,7 +5,7 @@
 -- Stability   : experimental
 -- Portability : POSIX
 --
--- This module exposes termonad's basic configuration options, as well as 'defaultMain', 'startWithCliArgs', and 'start'.
+-- This module exposes Termonad's basic configuration options, as well as 'defaultMain', 'startWithCliArgs', and 'start'.
 --
 -- If you want to configure Termonad, please take a look at "Termonad.Config".
 
@@ -16,5 +16,6 @@ module Termonad
   , module Config
   ) where
 
-import Termonad.App (defaultMain, start, startWithCliArgs)
+import Termonad.App (start)
 import Termonad.Config as Config
+import Termonad.Startup (defaultMain, startWithCliArgs)
